use RHEL Datasource into test_online_resize.py test
